+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Location NER confidence reduced
+    - LocationNER is less confident if lowecase is enabled
+    
 ## [0.4.1]  - 2020-03-22
 
 ### Added